safari css fix (#13)
on safari 11+ the title's background texts were displaced. fixed by using a media selector and a supports selector to only target safari 11+. the rest is just readjusting the numbers and in safari '\A' is required to put "get" and "involved" on separate lines. also fixed another issue where the navbar's links were too small. safari treated the maxsize as a transition from a minsize, so i moved the maxsize line above the transition line which resolved the issue.